Organized files
part one refactoring
