+++ robot_rumble/Characters/player.py
@@ -1,5 +1,4 @@
 import arcade
-import robot_rumble.constants as constants
 from importlib.resources import files
 
 # Character scaling constant
