+++ robot_rumble/Util/constants.py
@@ -14,6 +14,8 @@ JUMP_SPEED = 10
 RIGHT_FACING = 0
 LEFT_FACING = 1
 
+PLAYER_BULLET_MOVEMENT_SPEED = 0.4
+
 BULLET_SIZE = 2
 BULLET_SPEED = 8
 BULLET_RADIUS = 100
@@ -21,3 +23,9 @@ FORM_TIMER = 10
 
 BOSS_STUN_TIME = 3
 BOSS_PATH = [[530,530], [145,700], [915,700]]
+
+
+#Drone constants
+DRONE_MOVEMENT_SPEED = 0.25
+DRONE_TIMER = 0.2
+DRONE_BULLET_MOVEMENT_SPEED = 0.4
